+++ sampledb/frontend/objects/permissions.py
@@ -13,7 +13,7 @@ from flask_babel import _
 from ...logic.federation.update import update_poke_component
 from .. import frontend
 from ... import logic
-from ...logic import user_log
+from ...logic import user_log, background_tasks
 from ...logic.actions import get_action
 from ...logic.caching import cache_per_request
 from ...logic.components import get_component_by_uuid
@@ -216,6 +216,7 @@ def update_object_permissions(object_id: int) -> FlaskResponseT:
         if copy_permissions_form.validate_on_submit():
             logic.object_permissions.copy_permissions(object_id, int(copy_permissions_form.object_id.data))
             logic.object_permissions.set_user_object_permissions(object_id, flask_login.current_user.id, Permissions.GRANT)
+            background_tasks.post_trigger_object_permissions_webhooks(object_id)
             flask.flash(_("Successfully copied object permissions."), 'success')
     elif 'add_component_policy' in flask.request.form and add_component_policy_form.validate_on_submit():
         component_id = add_component_policy_form.component_id.data
@@ -275,6 +276,7 @@ def update_object_permissions(object_id: int) -> FlaskResponseT:
             flask.flash(_('No valid authentication method configured for %(component_name)s (%(component_address)s).', component_name=component.get_name(), component_address=component.address), 'warning')
         except requests.ConnectionError:
             flask.flash(_('Unable to contact %(component_name)s (%(component_address)s).', component_name=component.get_name(), component_address=component.address), 'warning')
+        background_tasks.post_trigger_object_permissions_webhooks(object_id)
         flask.flash(_("Successfully updated object permissions."), 'success')
     elif 'edit_component_policy' in flask.request.form and edit_component_policy_form.validate_on_submit():
         component_id = edit_component_policy_form.component_id.data
@@ -334,6 +336,7 @@ def update_object_permissions(object_id: int) -> FlaskResponseT:
             flask.flash(_('No valid authentication method configured for %(component_name)s (%(component_address)s).', component_name=component.get_name(), component_address=component.address), 'warning')
         except requests.ConnectionError:
             flask.flash(_('Unable to contact %(component_name)s (%(component_address)s).', component_name=component.get_name(), component_address=component.address), 'warning')
+        background_tasks.post_trigger_object_permissions_webhooks(object_id)
         flask.flash(_("Successfully updated object permissions."), 'success')
     else:
         if handle_permission_forms(
@@ -345,6 +348,7 @@ def update_object_permissions(object_id: int) -> FlaskResponseT:
             permissions_form
         ):
             user_log.edit_object_permissions(user_id=flask_login.current_user.id, object_id=object_id)
+            background_tasks.post_trigger_object_permissions_webhooks(object_id)
             flask.flash(_("Successfully updated object permissions."), 'success')
         else:
             flask.flash(_("A problem occurred while changing the object permissions. Please try again."), 'error')